+++ packages/recordScreen/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @u-moni/recordscreen
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @u-moni/common@1.1.5
+  - @u-moni/types@1.0.3
+
 ## 2.1.5
 
 ### Patch Changes